turning time from int to double

+++ C_RPi_Andy.cpp
@@ -38,7 +38,7 @@ int main(int argc, char* argv[])
     vector<float> azData;
 
     while (true) {
-     int time = sole->getRunningTime();
+     double time = sole->getRunningTime();
       cout << "\ntime: " << time << endl;
       // printf("time: %0.3f seconds \n", time);
       cout << "Cycle: " << cycle << endl;
